chore: migrate package manager from yarn 4 to pnpm 10
Setup
- Replace .yarnrc.yml with pnpm-lock.yaml; bump packageManager to pnpm@10.28.2.
- Convert yarn `resolutions` to `pnpm.overrides` (drop the `@npm:` prefix).
- Update CI workflows to use pnpm/action-setup with cache: pnpm.
- Update AGENTS.md commands and PR checklist.

Phantom-dependency fixes (pnpm is strict about declared deps)
- Add `long` and `@jest/globals` to devDependencies — they were resolved
  via yarn's hoisting before; pnpm requires explicit declaration.
- Convert two `import Long from 'long'` to `import type Long` in src
  (Long is only used as a type there).
- Fix index access in `getAudioWaveform` (rawData[i]! under
  noUncheckedIndexedAccess).

Lint cleanup applied by `eslint --fix` (prefer-optional-chain,
unnecessary-type-assertion). No behavior change.

Test runner
- Update `test`/`test:e2e` scripts to call `node_modules/jest/bin/jest.js`
  directly, since pnpm's `.bin` wrappers are sh scripts and can't be
  invoked through `node`.

+++ src/Utils/chat-utils.ts
@@ -387,7 +387,7 @@ export const extractSyncdPatches = async (result: BinaryNode, options: RequestIn
 						content = Buffer.from(Object.values(content))
 					}
 
-					const syncd = proto.SyncdPatch.decode(content as Uint8Array)
+					const syncd = proto.SyncdPatch.decode(content)
 					if (!syncd.version) {
 						syncd.version = { version: +collectionNode.attrs.version! + 1 }
 					}
